# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for مرشد الكتب الذكي - AI weather-based book recommender
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goodreads/shelf:
    get:
      operationId: getGoodreadsShelf
      tags: [goodreads]
      summary: Fetch a Goodreads user shelf
      description: Fetches books from a public Goodreads shelf via RSS (read or to-read)
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{1,20}$'
            minLength: 1
            maxLength: 20
          description: Goodreads numeric user ID
        - name: shelf
          in: query
          required: false
          schema:
            type: string
            enum: [read, to-read, currently-reading]
            default: read
          description: Which shelf to fetch
      responses:
        "200":
          description: Books on the shelf
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoodreadsShelf"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Failed to fetch from Goodreads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    GoodreadsBook:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        goodreadsId:
          type: string
        coverUrl:
          type: string
        rating:
          type: string
      required:
        - title
        - author
        - goodreadsId
    GoodreadsGenrePreference:
      type: object
      properties:
        genre:
          type: string
        score:
          type: number
      required:
        - genre
        - score
    GoodreadsShelf:
      type: object
      properties:
        userId:
          type: string
        shelf:
          type: string
        books:
          type: array
          items:
            $ref: "#/components/schemas/GoodreadsBook"
        genreProfile:
          type: array
          description: Genre signals inferred from public shelf titles and descriptions, ordered by strength.
          items:
            $ref: "#/components/schemas/GoodreadsGenrePreference"
        total:
          type: integer
          description: Reported shelf total when available, otherwise the best safe estimate.
        loaded:
          type: integer
          description: Number of unique books successfully loaded.
        reportedTotal:
          type: [integer, "null"]
          description: Total reported by Goodreads, when present in its RSS response.
        estimatedTotal:
          type: integer
          description: Best available total estimate, never less than loaded.
        complete:
          type: boolean
          description: Whether every available RSS page was fetched.
        warning:
          type: [string, "null"]
          description: User-facing explanation when the import is partial.
      required:
        - userId
        - shelf
        - books
        - genreProfile
        - total
        - loaded
        - reportedTotal
        - estimatedTotal
        - complete
        - warning
...